int -> size_t   + cherrypick

+++ Mesh_3/include/CGAL/Mesh_3/Worksharing_data_structures.h
@@ -844,7 +844,7 @@ protected:
     enqueue_task(create_task(wb, parent_task), parent_task);
   }
 
-  const int                         NUM_WORK_ITEMS_PER_BATCH;
+  const size_t                      NUM_WORK_ITEMS_PER_BATCH;
   TLS_WorkBuffer                    m_tls_work_buffers;
 };
